Move source files to src folder

+++ build.sh
@@ -1,9 +1,33 @@
-pyinstaller build.spec
+#!/bin/bash
+echo "*** Building Binary"
+binary=$(pyinstaller build.spec)
 
-rm -rf dist/plugins
-mkdir dist/plugins
-cp -R plugins/* dist/plugins/
-rm -rf dist/plugins/__init__.py* dist/plugins/__pycache__/
-rm dist/plugins/*/*.pyc 2>&1
-rm -rf dist/plugins/*/__pycache__/ 2>&1
+if [ $? -eq 0 ]; then
+	echo
+	echo "*** Copying plugins..."
+	rm -rf dist/plugins
+	mkdir dist/plugins
+	cp -R src/plugins dist/
 
+	echo
+	echo "*** Cleaning up..."
+	rm -rf dist/plugins/__init__.py* dist/plugins/__pycache__/
+	rm -rf dist/plugins/*/*.pyc
+	rm -rf dist/plugins/*/__pycache__/
+	rm -rf dist/plugins/GoogleApps/credentials.dat
+
+	echo
+	echo "*** Zipping..."
+	cd dist
+	zip ${binary%.*}.zip $binary plugins/
+	cd ..
+	
+	echo
+	echo "*** Done!"
+	exit 0
+else
+	echo
+	echo "*** Build failed :("
+	
+	exit 1
+fi
